requirements.txt
rich
numpy
PySide6
opencv-python

# Machine Learning Dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
